Moved SearchReducer into it own file
