+++ client/app/routes.js
@@ -2,7 +2,6 @@ import React from 'react';
 import {render} from 'react-dom';
 import App from './containers/App.jsx';
 import Index from './components/Index.jsx';
-import { searchReducer } from './components/SearchBar.jsx';
 import { Login, loginReducer } from './components/Login.jsx';
 import { Product, productReducer } from './components/Product.jsx';
 import NotFound from './components/NotFound.jsx';
@@ -11,6 +10,7 @@ import { syncHistoryWithStore, routerReducer, routerActions, routerMiddleware } 
 import { Provider } from 'react-redux';
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import { UserAuthWrapper} from 'redux-auth-wrapper';
+import { searchReducer } from './reducers/SearchReducer.js';
 
 
 const reducers = combineReducers({login: loginReducer, product: productReducer, search: searchReducer, routing:routerReducer});
